Female model with cyber armor (info in description))
The best way to handle equipping armor will be to have all of the sets on the player character at all times and enable/disable the game objects when switching. I already have the player set up to work like this. Doing it any other way might mess up the animations when I add them.