Started Working On Boss, And Slimes...
I think that you could add intensity to the emission (in the material) to the colors if you want them to stand out more, which I think they should, but it would then change the color of the original skin/hair so I thought you might want to do it. Also, the slimes need additional animation for the inner core body unless you wanted to just make that a sphere. Also feel free to mess with the slime material.